+++ src/common/enums/sms-service-name.enum.ts
@@ -1,0 +1,5 @@
+export enum SmsServiceName {
+  Kavenegar = 'kavenegar',
+  SmsIr = 'sms.ir',
+  Custom = 'custom_service_that_must_be_implemented_on_demand',
+}