+++ src/common/enums/socket-priority-level.enum.ts
@@ -1,4 +1,4 @@
-export enum SocketLevel {
+export enum SocketPriorityLevel {
   Info = 'info',
   Warn = 'warn',
   Error = 'error',