+++ src/common/models/notifications/subs/configs/sms-configs.model.ts
@@ -1,4 +1,7 @@
+import { SmsServiceName } from '../../../../enums';
+
 export class SmsConfigs {
+  service_name?: SmsServiceName;
   api_url!: string;
   api_key!: string;
 }