+++ src/common/models/notifications/subs/configs/socket-configs.model.ts
@@ -1,1 +1,7 @@
-export class SocketConfigs {}
+import { SocketServiceName } from '../../../../enums';
+
+export class SocketConfigs {
+  service_name?: SocketServiceName;
+  internal?: boolean;
+  configuration?: any;
+}